requirements.txt
Flask>=1.0.0
cloudant==2.4.0
requests_async>=0.6.2
flask_wtf>=0.14.3
slack_bolt>=1.0.1